Fix a warning in newer versions of pyyaml (#13)
Starting with pyyaml 5.1 the load function emits a warning on use. We're technically safe from
the issue that prompted this function being deprecated, but we're also not using any of the
features that doing anything other than safe_load gives us. So this just changes the call
to load() with safe_load().